fix & refactor: 로직을 추가하고 리팩토링함
