+++ src/components/public/bottomFloatingBar/BottomFloatingBar.test.tsx
@@ -1,65 +1,128 @@
+import { useRouter } from "next/navigation"
+
 import React from "react"
 
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import "@testing-library/jest-dom"
 import { fireEvent, render, screen } from "@testing-library/react"
 
 import BottomFloatingBar from "./BottomFloatingBar"
 
-describe("BottomFloatingBar", () => {
-  it("호스트가 아닐 때 올바르게 렌더링됩니다", () => {
-    render(<BottomFloatingBar isHost={false} />)
+// Next.js의 useRouter를 모킹합니다.
+jest.mock("next/navigation", () => {
+  return {
+    useRouter: jest.fn(),
+  }
+})
 
-    expect(screen.getByText("더 건강한 나와 팀을 위한 프로그램 🏃‍️️")).toBeInTheDocument()
-    expect(
-      screen.getByText("국내 최고 웰니스 전문가와 프로그램을 통해 지친 몸과 마음을 회복해봐요"),
-    ).toBeInTheDocument()
+// checkLogin 함수를 모킹합니다.
+jest.mock("@/actions/Auths/checkLogin", () => {
+  return {
+    __esModule: true,
+    default: jest.fn(() => {
+      return Promise.resolve(true)
+    }),
+  }
+})
+
+// navigator.clipboard를 모킹합니다.
+Object.assign(navigator, {
+  clipboard: {
+    writeText: jest.fn(),
+  },
+})
+
+describe("BottomFloatingBar", () => {
+  const mockSetHeight = jest.fn()
+  const mockRouter = { push: jest.fn(), replace: jest.fn() }
+  const queryClient = new QueryClient()
+
+  beforeEach(() => {
+    ;(useRouter as jest.Mock).mockReturnValue(mockRouter)
+  })
+
+  it("호스트가 아니고 참여하지 않았을 때 올바르게 렌더링됩니다", () => {
+    render(
+      <QueryClientProvider client={queryClient}>
+        <BottomFloatingBar
+          id="1"
+          isHost={false}
+          isJoined={false}
+          limit={10}
+          participant={3}
+          setHeight={mockSetHeight}
+        />
+      </QueryClientProvider>,
+    )
+
+    expect(screen.getByText("지금 모임에 참여하세요.")).toBeInTheDocument()
+    expect(screen.getByText("모임 개설확정까지 2명 남았어요!")).toBeInTheDocument()
     expect(screen.getByText("참여하기")).toBeInTheDocument()
-    expect(screen.queryByText("취소하기")).not.toBeInTheDocument()
+    expect(screen.getByText("공유하기")).toBeInTheDocument()
+    expect(screen.queryByText("개설취소")).not.toBeInTheDocument()
   })
 
   it("호스트일 때 올바르게 렌더링됩니다", () => {
-    render(<BottomFloatingBar isHost />)
+    render(
+      <QueryClientProvider client={queryClient}>
+        <BottomFloatingBar
+          id="1"
+          isHost
+          isJoined
+          limit={10}
+          participant={6}
+          setHeight={mockSetHeight}
+        />
+      </QueryClientProvider>,
+    )
 
-    expect(screen.getByText("더 건강한 나와 팀을 위한 프로그램 🏃‍️️")).toBeInTheDocument()
     expect(
       screen.getByText("모임을 공유해서 더 많은 사람들이 참여할 수 있도록 독려해봐요"),
     ).toBeInTheDocument()
-    expect(screen.getByText("공유하기")).toBeInTheDocument()
+    expect(screen.getByText("모집 정원 마감까지 4명 남았어요!")).toBeInTheDocument()
+    expect(screen.getByText("개설취소")).toBeInTheDocument()
     expect(screen.getByText("취소하기")).toBeInTheDocument()
+    expect(screen.getByText("공유하기")).toBeInTheDocument()
   })
 
   it("참여하기 버튼이 클릭 가능합니다", () => {
-    const mockOnClickParticipate = jest.fn()
-    jest.spyOn(React, "useState").mockImplementation(() => {
-      return [false, jest.fn()]
-    })
-    jest.spyOn(React, "useRef").mockReturnValue({ current: null })
-
-    render(<BottomFloatingBar isHost={false} />)
+    render(
+      <QueryClientProvider client={queryClient}>
+        <BottomFloatingBar
+          id="1"
+          isHost={false}
+          isJoined={false}
+          limit={10}
+          participant={3}
+          setHeight={mockSetHeight}
+        />
+      </QueryClientProvider>,
+    )
 
     const participateButton = screen.getByText("참여하기")
     fireEvent.click(participateButton)
 
-    expect(mockOnClickParticipate).toHaveBeenCalledTimes(0)
+    expect(participateButton).toBeInTheDocument()
   })
 
-  it("공유하기와 취소하기 버튼이 클릭 가능합니다", () => {
-    const mockOnClickShare = jest.fn()
-    const mockOnClickCancel = jest.fn()
-    jest.spyOn(React, "useState").mockImplementation(() => {
-      return [false, jest.fn()]
-    })
-    jest.spyOn(React, "useRef").mockReturnValue({ current: null })
-
-    render(<BottomFloatingBar isHost />)
+  it("공유하기 버튼이 클릭 가능합니다", () => {
+    render(
+      <QueryClientProvider client={queryClient}>
+        <BottomFloatingBar
+          id="1"
+          isHost={false}
+          isJoined={false}
+          limit={10}
+          participant={3}
+          setHeight={mockSetHeight}
+        />
+      </QueryClientProvider>,
+    )
 
     const shareButton = screen.getByText("공유하기")
-    const cancelButton = screen.getByText("취소하기")
-
     fireEvent.click(shareButton)
-    fireEvent.click(cancelButton)
 
-    expect(mockOnClickShare).toHaveBeenCalledTimes(0)
-    expect(mockOnClickCancel).toHaveBeenCalledTimes(0)
+    expect(shareButton).toBeInTheDocument()
+    expect(navigator.clipboard.writeText).toHaveBeenCalled()
   })
 })
